giving up on lifetime stuff for now. Made it work with basic, basic stuff
